optimize socket

+++ src/main/kotlin/com/oneliang/ktx/frame/socket/TcpPacketProcessor.kt
@@ -42,6 +42,6 @@ class TcpPacketProcessor(private val typeByteArrayLength: Int = 4, private val b
     fun receiveTcpPacket(inputStream: InputStream): TcpPacket {
         val type = receiveType(inputStream)
         val bodyByteArray = this.receiveBody(inputStream)
-        return TcpPacket(type.toInt(), bodyByteArray)
+        return TcpPacket(type, bodyByteArray)
     }
 }
